# openapi-projects.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ORCA CRM - Projects API",
    "description": "REST API endpoints for managing real estate projects, planning phases, unit inventory, construction progress reports, bookings, and integrations with the Accounting microservice.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/documents": {
      "get": {
        "summary": "Get document references from central registry",
        "parameters": [
          {
            "name": "entity",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "enum": ["project", "unit", "contract"] }
          },
          {
            "name": "entityId",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY"
      }
    }
  }
}